+++ mobile/app/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/mobile/di/ApplicationModule.kt
@@ -1,5 +1,7 @@
 package co.kr.parkjonghun.whatishedoingwithandroid.mobile.di
 
+import co.kr.parkjonghun.whatishedoingwithandroid.data.datasource.SampleDataSource
+import co.kr.parkjonghun.whatishedoingwithandroid.data.repository.SampleRepositoryImpl
 import co.kr.parkjonghun.whatishedoingwithandroid.domain.gateway.repository.SampleRepository
 import co.kr.parkjonghun.whatishedoingwithandroid.domain.usecase.SampleUseCase
 import co.kr.parkjonghun.whatishedoingwithandroid.domain.usecase.SampleUseCaseImpl
@@ -10,10 +12,30 @@ import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import javax.inject.Singleton
 
+// FIXME
 @Module
 @InstallIn(SingletonComponent::class, ViewModelComponent::class)
-class UseCaseModule {
+class ApplicationModule {
+    @Provides
+    @Singleton
+    fun provideSampleDataSource(): SampleDataSource {
+        return SampleDataSource()
+    }
+
+    @Provides
+    fun provideSampleRepository(
+        dataSource: SampleDataSource
+    ): SampleRepository {
+        val scope = CoroutineScope(Dispatchers.Main.immediate)
+
+        return SampleRepositoryImpl(
+            dataSource = dataSource,
+            scope = scope
+        )
+    }
+
     @Provides
     fun provideSampleUseCase(
         repository: SampleRepository,
